+++ bundlers/TypeScriptHelper.ts
@@ -48,7 +48,7 @@ module TypeScriptHelper {
     /** Read the 'prelude.js' source string used by browser-pack, insert the TypeScript static helpers (required for 'extends', annotations, and other TypeScript features) into it, and return it
      * @param [includeUsageComment=false] whether to include a comment in the source string explaining why the TypeScript static helpers are inserted
      */
-    export function insertTypeScriptHelpers(includeUsageComment?: boolean): Q.Promise<{ prelude: string }> {
+    export function createPreludeStringWithTypeScriptHelpers(includeUsageComment?: boolean): Q.Promise<{ prelude: string }> {
         return BundlifyHelper.getPreludeJsSource().then(function (preludeSrc) {
             var comment = "/* TypeScript static helpers - inserted once, here.  Run TypeScript compiler with '--noEmitHelpers' option to prevent duplicate helpers being inserted into each bundled TypeScript file */";