+++ utils/PathUtil.js
@@ -1,5 +1,5 @@
 "use strict";
-var gutil = require("gulp-util");
+var process = require("process");
 var StringUtil = require("./StringUtil");
 /** File system path utilities (some overlap with Node 'path' module).
  * Also contains a static 'projectRoot' property which can be initialized once and some of the other methods in this module can subsequently be called with relative paths
@@ -17,8 +17,8 @@ var PathUtil;
         return projectRoot;
     }
     PathUtil.getProjectRoot = getProjectRoot;
-    function getSetOrDefaultProjectPath(projRoot) {
-        projRoot = projRoot ? projRoot.replace(/\\/g, '/') : (projectRoot || (projectRoot = process.cwd().replace(/\\/g, '/')));
+    function getSetOrDefaultProjectPath(projRoot, projRootNormalized) {
+        projRoot = projRoot ? (projRootNormalized ? projRoot : projRoot.replace(/\\/g, '/')) : (projectRoot || (projectRoot = process.cwd().replace(/\\/g, '/')));
         return projRoot;
     }
     PathUtil.getSetOrDefaultProjectPath = getSetOrDefaultProjectPath;
@@ -28,22 +28,22 @@ var PathUtil;
      * @param file the file name to relativize
      * @param [projRoot] the optional path used to relativize the 'file', default value is 'process.cwd()'
      */
-    function toShortFileName(file, projRoot) {
-        projRoot = getSetOrDefaultProjectPath(projRoot);
+    function toShortFileName(file, projRoot, projRootNormalized) {
+        projRoot = getSetOrDefaultProjectPath(projRoot, projRootNormalized);
         var parts = file.replace(/\\/g, '/').split(projRoot);
-        return StringUtil.removeLeading(parts[parts.length - 1], '/');
+        return parts.length > 1 ? StringUtil.removeLeading(parts[parts.length - 1], '/') : file;
     }
     PathUtil.toShortFileName = toShortFileName;
     /** Create a custom proxy for a RegExp's test() function which logs messages to gulp-util each time test() is called
      * @param regex the regular expression to modify
      * @param show an object containing a boolean property 'showRegexTests' which is dynamically checked each time the RegExp's test() function is called to determine whether or not to print a message
      */
-    function createRegexInspector(regex, show) {
+    function createRegexInspector(regex, show, cb) {
         var origTest = regex.test;
         regex.test = function testInspector(str) {
             var res = origTest.call(regex, str);
             if (show.showRegexTests) {
-                gutil.log((res ? "building: " : "ignore: ") + toShortFileName(str));
+                cb(str, res);
             }
             return res;
         };